fix: 批量还原所有实体类为 Double，彻底消除冲突

+++ backend/src/main/java/com/stmanagement/dto/AttendanceDTO.java
@@ -6,7 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -18,11 +17,11 @@ public class AttendanceDTO {
     @NotNull private Long employeeId;
     private String employeeName;
     @NotNull private LocalDate workDate;
-    @NotNull private BigDecimal workHours;
-    private BigDecimal overtimeHours;
+    @NotNull private Double workHours;
+    private Double overtimeHours;
     private String clockIn;
     private String clockOut;
-    private BigDecimal totalHours;
+    private Double totalHours;
     private String workType;
     private String status;
     private String remark;
